Separate and comment code (#537)
